# backend/docs/swagger/swagger.yaml
basePath: /
...
host: localhost:8180
info:
  contact:
    email: bamort.support@trokan.de
    name: BaMoRT Support
  description: |-
    BaMoRT (MOAM Replacement) - Role-playing Game Character Management System API

    This API provides comprehensive character management, import/export functionality,
    master data management, and PDF generation for tabletop role-playing games.

    ## Authentication
    Most endpoints require authentication via JWT token. Obtain a token by calling POST /login.
    Include the token in the Authorization header: `Authorization: Bearer <token>`

    ## Character Import/Export
    The API supports pluggable adapters for importing/exporting characters from various systems.
    Supported formats include MOAM VTT, Foundry VTT, and more via microservice adapters.
  license:
    name: Custom License
    url: https://github.com/Bardioc26/bamort/blob/main/LICENSE
  title: BaMoRT API
  version: "2.0"
paths:
...
  /api/pdf/cleanup:
    post:
      description: Removes PDF files older than 1 hour from export_temp directory
      produces:
      - application/json
      responses:
        "200":
          description: Cleanup result with number of files deleted
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cleanup old PDF files
      tags:
      - PDF Export
  /api/pdf/export/{id}:
    get:
      description: Generates a PDF character sheet for the specified character
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        type: integer
      - description: 'Template name (default: Default_A4_Quer)'
        in: query
        name: template
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: PDF export result with filename and URL
          schema:
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Character not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Export character to PDF
      tags:
      - PDF Export
  /api/pdf/file/{filename}:
    get:
      description: Downloads a generated PDF file (public endpoint)
      parameters:
      - description: PDF filename
        in: path
        name: filename
        required: true
        type: string
      produces:
      - application/pdf
      responses:
        "200":
          description: PDF file
          schema:
            type: file
        "404":
          description: File not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get PDF file
      tags:
      - PDF Export
  /api/pdf/templates:
    get:
      description: Returns list of available PDF character sheet templates
      produces:
      - application/json
      responses:
        "200":
          description: List of template names
          schema:
            items:
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List PDF templates
      tags:
      - PDF Export
...
  /api/transfer/database/export:
    post:
      description: Exports the entire database as JSON (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: Database export data
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - admin access required
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Export full database
      tags:
      - Import/Export
  /api/transfer/database/import:
    post:
      consumes:
      - application/json
      description: Imports an entire database from JSON (admin only)
      parameters:
      - description: Database JSON data
        in: body
        name: database
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Import result
          schema:
            type: object
        "400":
          description: Invalid JSON data
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - admin access required
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Import full database
      tags:
      - Import/Export
  /api/transfer/download/{id}:
    get:
      description: Downloads a character as a JSON file attachment
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: JSON file download
          schema:
            type: file
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Character not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Download character as JSON file
      tags:
      - Import/Export
  /api/transfer/export/{id}:
    get:
      description: Exports a character as JSON for API consumption
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Character data in JSON format
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Character not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Export character as JSON
      tags:
      - Import/Export
  /api/transfer/import:
    post:
      consumes:
      - application/json
      description: Imports a character from JSON data
      parameters:
      - description: Character JSON data
        in: body
        name: character
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Imported character
          schema:
            type: object
        "400":
          description: Invalid JSON data
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Import character from JSON
      tags:
      - Import/Export
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"